Consolidate index membership, column-exists, and default-index checks
- Add _raise_if.index_values_in_stored_data for insert_rows and update
- Add _raise_if.cols_already_in_table mirroring cols_not_in_table
- Move has_still_default_index from insert_rows to common
